requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
langchain
unstructured
openai
chromadb
llama-index
langchain-google-genai
langchain-community